Fix: #55 Create a new stage phase for dense point cloud reconstruction (#71)
